feat: make ataxx a separate package
