+++ Cargo.toml
@@ -1,12 +1,4 @@
-[package]
-name = "mexx"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-num-traits = "0.2"
-num-derive = "0.4"
-strum = "0.26"
-strum_macros = "0.26"
+[workspace]
+members = [
+    "ataxx"
+]
