--- src/util/mod.rs
@@ -1,1 +1,0 @@
-pub mod type_macros;
